refactor: share elo engine
