+++ backend/src/driver/update-driver-elo-ratings.js
@@ -2,9 +2,11 @@ import mongoose from 'mongoose'
 import { fileURLToPath } from 'url'
 import connectDB from '../config/db.js'
 import Driver from './driver-model.js'
-import { expectedScore } from '../rating/elo-utils.js'
+import {
+  processRace,
+  DEFAULT_K
+} from '../rating/elo-engine.js'
 
-const DEFAULT_K = 20
 const DEFAULT_RATING = 1400
 const MIN_RACES = 5
 
@@ -14,39 +16,6 @@ const ensureConnection = async () => {
   }
 }
 
-const processRace = (placements, ratings, k) => {
-  const ids = Object.keys(placements)
-  const deltas = {}
-  for (let i = 0; i < ids.length; i++) {
-    for (let j = i + 1; j < ids.length; j++) {
-      const idA = ids[i]
-      const idB = ids[j]
-      const placeA = placements[idA]
-      const placeB = placements[idB]
-      if (placeA == null || placeB == null) continue
-      const entryA = ratings.get(idA) || { rating: DEFAULT_RATING, races: 0 }
-      const entryB = ratings.get(idB) || { rating: DEFAULT_RATING, races: 0 }
-      const ratingA = entryA.rating
-      const ratingB = entryB.rating
-      let outcomeA = 0.5
-      if (placeA < placeB) outcomeA = 1
-      else if (placeA > placeB) outcomeA = 0
-      const expectedA = expectedScore(ratingA, ratingB)
-      const expectedB = 1 - expectedA
-      const outcomeB = 1 - outcomeA
-      const deltaA = k * (outcomeA - expectedA)
-      const deltaB = k * (outcomeB - expectedB)
-      deltas[idA] = (deltas[idA] || 0) + deltaA
-      deltas[idB] = (deltas[idB] || 0) + deltaB
-    }
-  }
-  for (const id of ids) {
-    const base = ratings.get(id) || { rating: DEFAULT_RATING, races: 0 }
-    base.rating += deltas[id] || 0
-    base.races += 1
-    ratings.set(id, base)
-  }
-}
 
 const updateDriverEloRatings = async (k = DEFAULT_K, { disconnect = false } = {}) => {
   await ensureConnection()
@@ -78,7 +47,11 @@ const updateDriverEloRatings = async (k = DEFAULT_K, { disconnect = false } = {}
   const ratings = new Map()
 
   for (const race of raceList) {
-    processRace(race.placements, ratings, k)
+    processRace(race.placements, ratings, {
+      k,
+      defaultRating: DEFAULT_RATING,
+      raceDate: race.date
+    })
   }
 
   const bulk = Driver.collection.initializeUnorderedBulkOp()
